reincarnation commit
`magit-topgit.el' is no longer part of the Magit repository but
its karma is good and so it is reborn as a repository of its own.

To help it remember its past life do this:

  git remote add magit git://github.com/magit/magit.git
  git config remote.magit.tagopt --no-tags
  git fetch magit
  git replace --graft <this> 0a951ac7c7f152387002f75ebde1516cc7a506cb

See git-replace(1) for more information.

While it was still part of the Magit repository `magit-topgit.el'
was edited by the following people:

     1  Dan LaManna
     1  Graham Clark
    86  Jonas Bernoulli
     1  Laurent Laffont
     2  Leonardo Etcheverry
     1  Marc Herbert
     1  Noam Postavsky
     5  Phil Jackson
     3  Pieter Praet
     1  Rémi Vanicat
     2  Rüdiger Sonderfeld
     1  Thomas Jost
     5  Yann Hodique