reincarnation commit
`magit-topgit.el' is no longer part of the Magit repository but
its karma is good and so it is reborn as a repository of its own.

To help it remember its past life do this:

  git remote add magit git://github.com/magit/magit.git
  git config remote.magit.tagopt --no-tags
  git fetch magit
  git replace --graft <this> 1176825ca145e35823000564d3b0bedccd334d05

See git-replace(1) for more information.

While it was still part of the Magit repository `magit-topgit.el'
was edited by the following people:

    38  Jonas Bernoulli
     7  Yann Hodique
     3  Peter J. Weisberg
     2  Rüdiger Sonderfeld
     1  Dave Abrahams
     1  Nathan Weizenbaum
     1  Nicolas Dudebout
     1  Phil Jackson
     1  Pieter Praet